Savage

+++ src/homework/first/Pell.java
@@ -2,6 +2,7 @@ package homework.first;
 
 import java.util.Scanner;
 
+
 public class Pell {
     public static void main(String[] args) throws Exception {
         try (Scanner scanner = new Scanner(System.in)) {
@@ -14,7 +15,7 @@ public class Pell {
         }
     }
 
-    public static int pellNumbersWithRecursion(int n) {
+    public static long pellNumbersWithRecursion(int n) {
         if (n == 0) {
             return 0;
         } else if (n == 1) {
@@ -24,15 +25,15 @@ public class Pell {
         }
     }
 
-    public static int pellNumbersDefault(int n) {
-        int firstState = 0;
-        int secondState = 1;
+    public static long pellNumbersDefault(int n) {
+        long firstState = 0;
+        long secondState = 1;
         if (n == 0) {
             return 0;
         } else if (n == 1) {
             return 1;
         } else {
-            int value;
+            long value;
             for (int i = 2; i < n+1; i++) {  // Добавляем +1 чтобы посчитать необходимый нам элемент
                 value = 2 * secondState + firstState;
                 firstState = secondState; // Храним только 2 последних значения для пересчета
